workshop 2
What the code does:
Connects to a predefined WiFi network.

Displays a welcome message: “Hello student – Group 9”.

Fetches temperature data for the upcoming 24 hours using the SMHI public weather forecast API.

Tries to plot the temperature values as a line graph on the TFT screen.

Current status:
Right now, the code successfully:

Connects to WiFi

Displays the welcome message on the screen

However:

It fails to show the temperature graph and instead displays "Ingen data" (No data).

This might be caused by incorrect parsing of the SMHI JSON response or issues with how the graph is drawn.